+++ js/open_picture.js
@@ -1,4 +1,4 @@
-
+import { generateComment } from './picture.js';
 export const openingBigPicture = () => {
   const bigPicture = document.querySelector('.big-picture');
   const closeBtn = document.querySelector('.big-picture__cancel');
@@ -22,5 +22,14 @@ export const openingBigPicture = () => {
       commentCount.classList.remove('hidden');
       commentLoader.classList.remove('hidden');
     }
+  });
+  const connect = document.querySelectorAll('.picture');
+  const picture = document.querySelector('.big-picture__img img');
+  connect.forEach((elem) => {
+    elem.addEventListener('click', (e) => {
+      picture.setAttribute('src', e.target.getAttribute('src'));
+      picture.setAttribute('alt', e.target.getAttribute('alt'));
+      generateComment();
+    });
   });
 };
